Add gas saving (#28)

+++ contracts/interfaces/IReClammPool.sol
@@ -10,7 +10,7 @@ struct ReClammPoolParams {
     string symbol;
     string version;
     uint256 increaseDayRate;
-    uint256 sqrtQ0;
+    uint96 sqrtQ0;
     uint256 centerednessMargin;
 }
 
@@ -21,7 +21,7 @@ interface IReClammPool is IBasePool {
     /// @dev The function is not implemented.
     error NotImplemented();
 
-    event SqrtQ0Updated(uint256 startSqrtQ0, uint256 endSqrtQ0, uint256 startTime, uint256 endTime);
+    event SqrtQ0Updated(uint96 startSqrtQ0, uint96 endSqrtQ0, uint32 startTime, uint32 endTime);
 
     event VirtualBalancesUpdated(uint256[] virtualBalances);
 
@@ -33,9 +33,9 @@ interface IReClammPool is IBasePool {
 
     function getLastTimestamp() external view returns (uint256);
 
-    function getCurrentSqrtQ0() external view returns (uint256);
+    function getCurrentSqrtQ0() external view returns (uint96);
 
-    function setSqrtQ0(uint256 newSqrtQ0, uint256 startTime, uint256 endTime) external;
+    function setSqrtQ0(uint96 newSqrtQ0, uint32 startTime, uint32 endTime) external;
 
     function setIncreaseDayRate(uint256 newIncreaseDayRate) external;
 }
